+++ examples/rtd_images/RTD_E12_matplotlib_integration.py
@@ -3,7 +3,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-from plotez import plot_errorbar, plot_xy
+from plotez import SAVE_DPI, plot_errorbar, plot_xy
 
 fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(12, 4))
 
@@ -19,5 +19,5 @@ y_err = 0.1
 plot_errorbar(x2, y2, y_err=y_err, x_label="X", data_label="cos(x)", axis=ax2)
 
 # plt.show()
-plt.savefig("RTD_E12_matplotlib_integration.png", dpi=300)
+plt.savefig("RTD_E12_matplotlib_integration.png", dpi=SAVE_DPI)
 plt.close()