+++ examples/rtd_images/RTD_E3_scatter_plot.py
@@ -3,7 +3,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-from plotez import plot_xy
+from plotez import SAVE_DPI, plot_xy
 
 # set a default generator for reproducibility
 rng = np.random.default_rng(1234)
@@ -15,5 +15,5 @@ y = 2 * x + 1 + rng.random(x.shape) * 0.2
 plot_xy(x, y, x_label="X", y_label="Y", plot_title="Noisy Relationship", is_scatter=True)
 
 # plt.show()
-plt.savefig("RTD_E3_scatter_plot.png", dpi=300)
+plt.savefig("RTD_E3_scatter_plot.png", dpi=SAVE_DPI)
 plt.close()