+++ examples/rtd_images/RTD_E6_asym_errors.py
@@ -3,7 +3,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-from plotez import plot_errorbar
+from plotez import SAVE_DPI, plot_errorbar
 
 x = np.linspace(0, 10, 20)
 y = x**2 / 10
@@ -13,5 +13,5 @@ y_err = np.array([0.1 * y, 0.35 * y])  # Lower (10% of value)  # Upper  (35% of 
 plot_errorbar(x, y, y_err=y_err)
 
 # plt.show()
-plt.savefig("RTD_E6_asym_errors.png", dpi=300)
+plt.savefig("RTD_E6_asym_errors.png", dpi=SAVE_DPI)
 plt.close()